chore: DealiColor 파일 분리
